% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fillna.R
\name{fillna}
\alias{fillna}
\title{A function that sets any existing object's NA values to a user specified
value (p).}
\usage{
fillna(x, p = 0)
}
\arguments{
\item{x}{a vector, matrix, data frame, a non-hierarchical list,
or a list of no more than 2 hierarchy levels (i.e. a list of lists)}

\item{p}{a value with which to replace the NA values. Defaults to 0.}
}
\value{
an object of the same type as x with any NA values set to p.
}
\description{
This function takes a variety of inputs and sets any of the input's
NA values to a user-specified value (p).
}
\examples{
fillna(matrix(c(NA, 7, NA, 6), 2, 2))
fillna(c(NA, 6))
fillna(list(NA, 3), -99)
fillna(list(NA, 3, list(4, NA)), -99)

mtcars[1,3] <- NA
fillna(mtcars, 0)
}

