% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interactions.R
\name{interactions}
\alias{interactions}
\title{A function that returns all pairwise interations between the input's
columns}
\usage{
interactions(x)
}
\arguments{
\item{x}{a matrix or dataframe with at least two columns}
}
\value{
an object of the same type as x with {p choose 2} additional
columns representing the pairwise interactions that are concatenated to the
original object x.
}
\description{
This function takes in a matrix or dataframe and returns a data structure
of the same type that includes the original columns concatenated with all
pairwise interactions.
}
\examples{
interactions(mtcars)
}

