% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/num_unique_values.R
\name{num_unique_values}
\alias{num_unique_values}
\title{A function to count the number of unique values in a vector, matrix,
or dataframe.}
\usage{
num_unique_values(x)
}
\arguments{
\item{x}{a vector, matrix, or dataframe.}
}
\value{
a 1x1 vector (if input is vector) or
as a px1 vector (if input is a matrix or dataframewith p columns).
}
\description{
This function takes in either a vector, matrix, or dataframe and outputs
the number of unique values either as a 1x1 vector (if input is vector) or
as a px1 vector (if input is a matrix or dataframewith p columns).
}

