% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qmatrix.R
\name{qmatrix}
\alias{qmatrix}
\title{A function to make an n x p matrix of randomly generated integers with
minimal effort by the user.}
\usage{
qmatrix(n, p, a = 0, b = 2 * p * n)
}
\arguments{
\item{n}{the number of rows for the matrix}

\item{p}{the number of columns for the matrix}

\item{a}{the lower bound of the discrete uniform distribution used to
simulate the matrix elements. Defaults to 0.}

\item{b}{the upper bound of the discrete uniform distribution used to
simulate the matrix elements. Defaults to 2*p*n}
}
\value{
a matrix if size n x p.
}
\description{
This function takes in the dimensions of the desired matrix and outputs
a matrix whose elements are randomly generated integers. The
dimensions are the only required input, but users can also control the
parameters of the discrete uniform used to generate the matrix.
}

