% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qvec.R
\name{qvec}
\alias{qvec}
\title{A function to make a p x 1 vector of randomly generated integers with
minimal effort by the user.}
\usage{
qvec(p, a = 0, b = 2 * p)
}
\arguments{
\item{p}{the number of elements of the vector}

\item{a}{the lower bound of the discrete uniform distribution used to
simulate the matrix elements. Defaults to 0.}

\item{b}{the upper bound of the discrete uniform distribution used to
simulate the matrix elements. Defaults to 2*p*n}
}
\value{
a vector if length p x 1.
}
\description{
This function takes in the length p of the desired vector and outputs
a vector whose elements are randomly generated integers. The
length is the only required input, but users can also control the
parameters of the discrete uniform used to generate the vector
}

