% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time_it.R
\name{time_it}
\alias{time_it}
\title{A function to print the run time of its input function call and save the
results}
\usage{
time_it(FUN)
}
\arguments{
\item{FUN}{call to a function}
}
\value{
the result of the inputted function call while printing the
elapsed time taken to call function.
}
\description{
This function takes a function call, prints the time taken to run the
function call, and returns the result of the function call.
}
\examples{
my_matrix <- matrix(rnorm(10000), 100, 100)
eigen_results <- time_it(eigen(my_matrix))
}

